酒店图片上传demo

+++ src/main/java/cc/mrbird/febs/system/entity/File.java
@@ -52,7 +52,7 @@ public class File {
      * 附件大小
      */
     @TableField("file_size")
-    private Double fileSize;
+    private Long fileSize;
 
     /**
      * 缩略图url
